+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Unreleased
 
+### Features
+
+-   Added a new `ParameterType.Object` for declaring object options which will be shallowly merged when read from user configuration.
+-   Added a new `Application.EVENT_BOOTSTRAP_END` event emitted when `Application.bootstrap` is called.
+
 ### Bug Fixes
 
 -   TypeDoc will now work properly in packages mode when converting packages outside the current working directory, #2043.
 -   Fixed deprecation warning for `isIdentifierOrPrivateIdentifier`.
 -   Fixed centering of title bar on wide displays, #2046.
+
+### Thanks!
+
+-   @citkane
 
 ## v0.23.11 (2022-08-26)